Bug fixes + New test files for Resources

+++ planner/src/main/java/edu/isi/wings/execution/logger/api/impl/kb/RunKB.java
@@ -265,7 +265,8 @@ public class RunKB implements ExecutionLoggerAPI, ExecutionMonitorAPI {
 
 	private void deleteGraph(String id) {
 	  try {
-      ontologyFactory.getKB(new URIEntity(id).getURL(), OntSpec.PLAIN).delete();
+	    if(id != null)
+	      ontologyFactory.getKB(new URIEntity(id).getURL(), OntSpec.PLAIN).delete();
     } catch (Exception e) {
       e.printStackTrace();
     }
